+++ src/controller.ts
@@ -6,8 +6,9 @@ import { existsSync } from 'fs';
 import { PythonShell } from 'python-shell';
 import * as modtool from './modtool';
 import * as blocks from './blockFilter';
+import { getBlockFilesTree } from './moduleTree';
 
-export class GNURadioController implements vscode.TreeDataProvider<vscode.TreeItem> {
+export class GNURadioController {
     private readonly extId: string;
     private readonly extRoot: vscode.Uri;
     private _outputChannel: vscode.OutputChannel;
@@ -341,7 +342,7 @@ export class GNURadioController implements vscode.TreeDataProvider<vscode.TreeIt
             if (existingBlocks.has(blockName)) {
                 warningMessage = `Are you sure you want to remove "${blockName}"?`;
                 successMessage = `Block "${blockName}" was removed`;
-                detailMessage = (await modtool.getBlockFilesTree(blockName, vscode.Uri.file(this.cwd!), this.moduleName!))
+                detailMessage = (await getBlockFilesTree(blockName, vscode.Uri.file(this.cwd!), this.moduleName!))
                     .map(item => item.resourceUri!.fsPath.slice(this.cwd!.length + 1));
                 detailMessage.unshift('The following files will be deleted:');
             } else {
@@ -403,7 +404,7 @@ export class GNURadioController implements vscode.TreeDataProvider<vscode.TreeIt
             if (!newBlockName) {
                 return;
             }
-            let blockFiles = (await modtool.getBlockFilesTree(blockName, vscode.Uri.file(this.cwd!), this.moduleName!))
+            let blockFiles = (await getBlockFilesTree(blockName, vscode.Uri.file(this.cwd!), this.moduleName!))
                 .map(item => item.resourceUri!.fsPath.slice(this.cwd!.length + 1));
             blockFiles.unshift('The following files will be renamed:');
             const confirm = await vscode.window.showWarningMessage(
@@ -516,52 +517,5 @@ export class GNURadioController implements vscode.TreeDataProvider<vscode.TreeIt
                 vscode.window.showErrorMessage(err.message);
             }
         }
-    }
-
-    public async getTreeItem(element: vscode.TreeItem) {
-        return element;
-    }
-
-    public async getChildren(element?: vscode.TreeItem) {
-        if (!this.cwd) {
-            return [];
-        }
-        if (!this.moduleName) {
-            await this.setCwd(this.cwd);
-        }
-        if (!this.moduleName) {
-            vscode.window.showInformationMessage('No GNURadio Module detected in the workspace');
-            return [];
-        }
-        if (element) {
-            if (!element.label) {
-                element.collapsibleState = vscode.TreeItemCollapsibleState.None;
-                return [];
-            }
-            const baseUri = vscode.Uri.file(this.cwd);
-            return await modtool.getBlockFilesTree(element.label.toString(), baseUri, this.moduleName);
-        } else {
-            const cppBlocks = blocks.getCppBlocks(this.cwd, this.moduleName);
-            const xmlBlocks = blocks.getXmlBlocks(this.cwd, this.moduleName);
-            return Array.from(blocks.getAllBlocks(this.cwd, this.moduleName))
-                .map((name) => {
-                    let item = new vscode.TreeItem(name, vscode.TreeItemCollapsibleState.Collapsed);
-                    item.contextValue = 'block';
-                    if (cppBlocks.includes(name)) {
-                        item.contextValue += '.cpp';
-                    }
-                    if (xmlBlocks.includes(name)) {
-                        item.contextValue += '.xml';
-                    }
-                    return item;
-                });
-        }
-    }
-
-    private _onDidChangeTreeData = new vscode.EventEmitter<vscode.TreeItem | undefined | null | void>();
-    readonly onDidChangeTreeData = this._onDidChangeTreeData.event;
-
-    public refresh() {
-        this._onDidChangeTreeData.fire();
     }
 }